move green buttons to separate component
