+++ src/components/mapart/mapSettings.css
@@ -5,7 +5,3 @@
 .preProcessingInputBox {
   width: 3em;
 }
-
-.donateA {
-  text-decoration: none;
-}
