+++ src/components/mapart/mapSettings.js
@@ -1,10 +1,8 @@
 import React, { Component } from "react";
-import { gzip } from "pako";
 
 import AutoCompleteInputBlockToAdd from "./autoCompleteInputBlockToAdd/autoCompleteInputBlockToAdd";
 
 import Tooltip from "../tooltip";
-import coloursJSON from "./coloursJSON.json";
 
 import BackgroundColourModes from "./json/backgroundColourModes.json";
 import DitherMethods from "./json/ditherMethods.json";
@@ -13,132 +11,10 @@ import StaircaseModes from "./json/staircaseModes.json";
 import SupportedVersions from "./json/supportedVersions.json";
 import WhereSupportBlocksModes from "./json/whereSupportBlocksModes.json";
 
-import NBTWorker from "./workers/nbt.jsworker";
-
 import "./mapSettings.css";
 
 class MapSettings extends Component {
-  state = {
-    buttonWidth_NBT_Joined: 1,
-    buttonWidth_NBT_Split: 1,
-    buttonWidth_Mapdat_Split: 1,
-    mapPreviewWorker_onFinishCallback: null,
-  };
-
-  nbtWorker = new Worker(NBTWorker);
-
-  resetButtonWidths() {
-    this.setState({ buttonWidth_NBT_Joined: 1, buttonWidth_NBT_Split: 1, buttonWidth_Mapdat_Split: 1 });
-  }
-
-  getNBT_base = (workerHeader) => {
-    const {
-      getLocaleString,
-      optionValue_version,
-      optionValue_staircasing,
-      optionValue_whereSupportBlocks,
-      optionValue_supportBlock,
-      uploadedImage_baseFilename,
-      currentMaterialsData,
-      downloadBlobFile,
-      mapPreviewWorker_inProgress,
-    } = this.props;
-    if (mapPreviewWorker_inProgress) {
-      this.setState({ mapPreviewWorker_onFinishCallback: () => this.getNBT_base(workerHeader) });
-      return;
-    }
-    if (Object.entries(currentMaterialsData.currentSelectedBlocks).every((elt) => elt[1] === "-1")) {
-      alert(getLocaleString("DOWNLOAD/ERROR-NONE-SELECTED"));
-      return;
-    }
-    this.nbtWorker.terminate();
-    this.resetButtonWidths();
-    const t0 = performance.now();
-    this.nbtWorker = new Worker(NBTWorker);
-    this.nbtWorker.onmessage = (e) => {
-      switch (e.data.head) {
-        case "PROGRESS_REPORT_NBT_JOINED": {
-          this.setState({ buttonWidth_NBT_Joined: e.data.body });
-          break;
-        }
-        case "PROGRESS_REPORT_NBT_SPLIT": {
-          this.setState({ buttonWidth_NBT_Split: e.data.body });
-          break;
-        }
-        case "PROGRESS_REPORT_MAPDAT_SPLIT": {
-          this.setState({ buttonWidth_Mapdat_Split: e.data.body });
-          break;
-        }
-        case "NBT_ARRAY": {
-          const t1 = performance.now();
-          console.log(`Created NBT by ${(t1 - t0).toString()}ms`);
-          const { NBT_Array, whichMap_x, whichMap_y } = e.data.body;
-          downloadBlobFile(
-            gzip(NBT_Array),
-            "application/x-minecraft-level",
-            workerHeader === "CREATE_NBT_SPLIT" ? `${uploadedImage_baseFilename}_${whichMap_x}_${whichMap_y}.nbt` : `${uploadedImage_baseFilename}.nbt`
-          );
-          break;
-        }
-        case "MAPDAT_BYTES": {
-          const t1 = performance.now();
-          console.log(`Created Mapdat by ${(t1 - t0).toString()}ms`);
-          const { Mapdat_Bytes, whichMap_x, whichMap_y } = e.data.body;
-          downloadBlobFile(gzip(Mapdat_Bytes), "application/x-minecraft-map", `${uploadedImage_baseFilename}_${whichMap_x}_${whichMap_y}.dat`);
-          break;
-        }
-        default: {
-          throw new Error("Unknown worker response header");
-        }
-      }
-    };
-    this.nbtWorker.postMessage({
-      head: workerHeader,
-      body: {
-        coloursJSON: coloursJSON,
-        StaircaseModes: StaircaseModes,
-        WhereSupportBlocksModes: WhereSupportBlocksModes,
-        optionValue_version: optionValue_version,
-        optionValue_staircasing: optionValue_staircasing,
-        optionValue_whereSupportBlocks: optionValue_whereSupportBlocks,
-        optionValue_supportBlock: optionValue_supportBlock,
-        pixelsData: currentMaterialsData.pixelsData,
-        maps: currentMaterialsData.maps,
-        currentSelectedBlocks: currentMaterialsData.currentSelectedBlocks,
-      },
-    });
-  };
-
-  onGetNBTClicked = () => {
-    this.getNBT_base("CREATE_NBT_JOINED");
-  };
-
-  onGetNBTSplitClicked = () => {
-    this.getNBT_base("CREATE_NBT_SPLIT");
-  };
-
-  onGetMapdatSplitClicked = () => {
-    this.getNBT_base("CREATE_MAPDAT_SPLIT");
-  };
-
-  componentDidUpdate(prevProps) {
-    const { optionValue_modeNBTOrMapdat, mapPreviewWorker_inProgress } = this.props;
-    if (prevProps.optionValue_modeNBTOrMapdat !== optionValue_modeNBTOrMapdat) {
-      // reset callback if changing mode from NBT to mapdat while rendering after download button clicked (very niche but a bug squashed nontheless)
-      this.setState({ mapPreviewWorker_onFinishCallback: null });
-    }
-    if (!mapPreviewWorker_inProgress && this.state.mapPreviewWorker_onFinishCallback !== null) {
-      this.state.mapPreviewWorker_onFinishCallback();
-      this.setState({ mapPreviewWorker_onFinishCallback: null });
-    }
-  }
-
-  componentWillUnmount() {
-    this.nbtWorker.terminate();
-  }
-
   render() {
-    const { buttonWidth_NBT_Joined, buttonWidth_NBT_Split, buttonWidth_Mapdat_Split } = this.state;
     const {
       getLocaleString,
       optionValue_version,
@@ -181,7 +57,6 @@ class MapSettings extends Component {
       onOptionChange_PreProcessingBackgroundColourSelect,
       preProcessingValue_backgroundColour,
       onOptionChange_PreProcessingBackgroundColour,
-      onViewOnlineClicked,
     } = this.props;
     const setting_mode = (
       <React.Fragment>
@@ -518,70 +393,6 @@ class MapSettings extends Component {
         <br />
       </React.Fragment>
     );
-    let buttons_mapModeConditional;
-    if (optionValue_modeNBTOrMapdat === MapModes.SCHEMATIC_NBT.uniqueId) {
-      buttons_mapModeConditional = (
-        <React.Fragment>
-          <Tooltip tooltipText={getLocaleString("VIEW-ONLINE/TITLE-TT")}>
-            <span className="greenButton_old" onClick={onViewOnlineClicked}>
-              {getLocaleString("VIEW-ONLINE/TITLE")}
-            </span>
-          </Tooltip>
-          <br />
-          <Tooltip tooltipText={getLocaleString("DOWNLOAD/NBT-SPECIFIC/DOWNLOAD-TT")}>
-            <div className="greenButton greenButton_large" style={{ display: "block" }} onClick={this.onGetNBTClicked}>
-              <span className="greenButton_text greenButton_large_text">{getLocaleString("DOWNLOAD/NBT-SPECIFIC/DOWNLOAD")}</span>
-              <div
-                className="greenButton_progressDiv"
-                style={{
-                  width: `${Math.floor(buttonWidth_NBT_Joined * 100)}%`,
-                }}
-              />
-            </div>
-          </Tooltip>
-          <br />
-          <Tooltip tooltipText={getLocaleString("DOWNLOAD/NBT-SPECIFIC/DOWNLOAD-SPLIT-TT")}>
-            <div className="greenButton" style={{ display: "block" }} onClick={this.onGetNBTSplitClicked}>
-              <span className="greenButton_text">{getLocaleString("DOWNLOAD/NBT-SPECIFIC/DOWNLOAD-SPLIT")}</span>
-              <div
-                className="greenButton_progressDiv"
-                style={{
-                  width: `${Math.floor(buttonWidth_NBT_Split * 100)}%`,
-                }}
-              />
-            </div>
-          </Tooltip>
-          <br />
-        </React.Fragment>
-      );
-    } else {
-      buttons_mapModeConditional = (
-        <React.Fragment>
-          <Tooltip tooltipText={getLocaleString("DOWNLOAD/MAPDAT-SPECIFIC/DOWNLOAD-TT")}>
-            <div className="greenButton greenButton_large" style={{ display: "block" }} onClick={this.onGetMapdatSplitClicked}>
-              <span className="greenButton_text greenButton_large_text">{getLocaleString("DOWNLOAD/MAPDAT-SPECIFIC/DOWNLOAD")}</span>
-              <div
-                className="greenButton_progressDiv"
-                style={{
-                  width: `${Math.floor(buttonWidth_Mapdat_Split * 100)}%`,
-                }}
-              />
-            </div>
-          </Tooltip>
-          <br />
-        </React.Fragment>
-      );
-    }
-    const button_donate = (
-      <React.Fragment>
-        <Tooltip tooltipText={getLocaleString("DONATE/TITLE-TT")}>
-          <a className="donateA" href="./supporters">
-            <span className="greenButton_old">{getLocaleString("DONATE/TITLE")}</span>
-          </a>
-        </Tooltip>
-        <br />
-      </React.Fragment>
-    );
     const settingsDiv = (
       <div className="section settingsDiv">
         <h2>{getLocaleString("MAP-SETTINGS/TITLE")}</h2>
@@ -595,8 +406,6 @@ class MapSettings extends Component {
         {setting_betterColour}
         {setting_dithering}
         {preprocessing}
-        {buttons_mapModeConditional}
-        {button_donate}
       </div>
     );
     return settingsDiv;
