+++ src/locale/zh-Hans/strings.json
@@ -129,7 +129,8 @@
     "MAPDAT-SPECIFIC": {
       "DOWNLOAD": "下载",
       "DOWNLOAD-TT": "下载map.dat分成1x1区块"
-    }
+    },
+    "ZIPPING": null
   },
   "MATERIALS": {
     "TITLE": "材料",
